pyproject.toml
[project]
name = "new-easys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "django-json-widget>=2.0.1",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "ipdb>=0.13.13",
    "ipython>=9.1.0",
    "ruff>=0.11.5",
]

[tool.ruff]
# Allow lines to be as long as 120.
line-length = 120